Move Web API mocks
